Updated and rewrote the code on the crash-test and index html
Crash-Test.HTML
Added some comments and used it to reference the layout of the index code.

Index.HTML
Erased and then recreated the code using a combo of the previously written Crash code and the course video lesson for this first web page.
Added comments, Added Copyright icon and Favicon to the site. Also, changed the friends images to links for more information on them.